LayoutTests/inspector/unit-tests/index-set-expected.txt

Tests for WI.IndexSet.


== Running test suite: IndexSet
-- Running test case: IndexSet.constructor
PASS: size should be zero.
PASS: firstIndex should be NaN.
PASS: lastIndex should be NaN.
PASS: Should be [].

-- Running test case: IndexSet.constructor array
Initialize IndexSet with an array.
PASS: size should be 5.
PASS: firstIndex should be 0.
PASS: lastIndex should be 100.
PASS: Should be [0,1,5,50,100].
Initialize IndexSet with an array containing duplicate indexes.
PASS: size should be 5.
PASS: Should be [0,1,5,50,100].

-- Running test case: IndexSet.constructor invalid
PASS: size should be zero.

-- Running test case: IndexSet.prototype.clear
PASS: size should be zero.
PASS: firstIndex should be NaN.
PASS: lastIndex should be NaN.
PASS: Should be [].

-- Running test case: IndexSet.prototype.add
PASS: size should be 1.
PASS: has should return true.

-- Running test case: IndexSet.prototype.add duplicate
PASS: size should be 1.

-- Running test case: IndexSet.prototype.add invalid
PASS: size should be zero.

-- Running test case: IndexSet.prototype.delete
Given an IndexSet with values [1,2,3]:
PASS: delete 3 should succeed.
PASS: has 3 should return false.

-- Running test case: IndexSet.prototype.delete nonexistent
Given an IndexSet with values [1,2,3]:
PASS: delete 4 should fail.

-- Running test case: IndexSet.prototype.delete invalid
-- Running test case: IndexSet.prototype[Symbol.iterator]
Given an IndexSet with values [20,1,10,2]:
1
2
10
20

-- Running test case: IndexSet.prototype.indexGreaterThan
Given an IndexSet with values [1,2]:
PASS: Index greater than 0 should be 1.
PASS: Index greater than 1 should be 2.
PASS: Index greater than 2 should be NaN.
PASS: Index greater than 3 should be NaN.

-- Running test case: IndexSet.prototype.indexLessThan
Given an IndexSet with values [1,2]:
PASS: Index less than 0 should be NaN.
PASS: Index less than 1 should be NaN.
PASS: Index less than 2 should be 1.
PASS: Index less than 3 should be 2.
